Introduce the `People` class
This class is in charge of building a `Person` object for each row in
the raw content for each format.

It is meant to easily support new formats as the only thing that needs
to be defined is the parser which needs to be added to the
`Parsers::LIST` constant.