Stop inheriting from `CSV`
I had been debating with the idea of inheriting directly from this class
from the stantard library. Initially I did because it felt like it made
each of the initial parser quite powerful, upto par with `CSV`. However,
given that we're not really using any other of its capabilities,
inheriting from it just made the resulting class a bit more complicated
that it needed to be.